requirements.txt
scipy==1.13.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
python-dotenv==1.0.1
pytest==8.2.0
matplotlib==3.9.0
mlflow 
scikit-learn
langchain 
langchain-community 
langchain-google-genai 
chromadb 
google-generativeai 
pypdf
langchain-nvidia-ai-endpoints
langchain-core
fastapi 
uvicorn 
strawberry-graphql 
python-jose[cryptography] 
tinydb